3816 Sidebar tests"
