+++ src/sidebar/ConnectedSidebar.module.scss
